Add Medium C++ Solution for 1725. Number of Sets of K Non-Overlapping Line Segments from LeetCode

+++ LeetCode/Patterns/Dynamic Programming/1725. Number of Sets of K Non-Overlapping Line Segments/Solution.cpp
@@ -6,8 +6,8 @@ public:
         if(k==0) return 1;
         if(i>=n) return 0;
         if(dp[k][i]!=-1) return dp[k][i]%mod;
-        int skip = solve(n,k,i+1)%mod;
-        int take=0;
+        long long skip = solve(n,k,i+1)%mod;
+        long long take=0;
         for(int j=i+1;j<n;j++){
             take=(take+solve(n,k-1,j))%mod;
         }
